refactor: remove presentational logic from transaction module
We remove presentational logic from `ui/src/transaction.ts` and move it
into the components that use it. This eliminates circular dependencies
involving the `transaction` module.
