+++ ui/DesignSystem/Wallet/Transactions/TxListItem.svelte
@@ -6,11 +6,37 @@
  LICENSE file.
 -->
 <script lang="typescript">
+  import type { SvelteComponent } from "svelte";
+  import { Icon } from "ui/DesignSystem";
   import dayjs from "dayjs";
-  import type { Tx } from "ui/src/transaction";
-  import { txIcon } from "ui/src/transaction";
+  import { Tx, TxKind } from "ui/src/transaction";
 
   export let tx: Tx;
+
+  function txIcon(tx: Tx): typeof SvelteComponent {
+    switch (tx.kind) {
+      case TxKind.ClaimRadicleIdentity:
+        return Icon.Registered;
+      case TxKind.CollectFunds:
+      case TxKind.Withdraw:
+        return Icon.Withdraw;
+      case TxKind.CommitEnsName:
+      case TxKind.Erc20Allowance:
+      case TxKind.LinkEnsNameToOrg:
+      case TxKind.RegisterEnsName:
+      case TxKind.UpdateEnsMetadata:
+        return Icon.Ethereum;
+      case TxKind.SupportOnboarding:
+      case TxKind.UpdateSupport:
+        return Icon.TokenStreams;
+      case TxKind.TopUp:
+        return Icon.Topup;
+      case TxKind.CreateOrg:
+        return Icon.Orgs;
+      case TxKind.AnchorProject:
+        return Icon.Anchor;
+    }
+  }
 </script>
 
 <style>
